Move upload limit tests into test_upload_books module
